added -n option, currencies are now double precision
README
	uninstallation now documented in UNINSTALLING
	-n option now documented

btcapi.h
	currencies are now represented by double

btcapi.c
	removal of #if DEBUG checks -> checked in btcdbg.c instead

btcdbg.c
	definition of IGNORE() macro, ignores unused parameters
	checks for DEBUG macro in btcdbg()

cmdlineutils.c
	-n option added to --help
	btcdbg added to resetw() and resetb()

main.c
	addition of n variable (double) for --amount
	removed redundant fwide() function call

+++ src/include/btcapi.h
@@ -34,11 +34,11 @@ typedef struct {
 // struct containing current exchange information
 
 typedef struct {
-	float buy;
+	double buy;
 	currcy_t currcy;
 	bool got;
 	bool result;
-	float sell;
+	double sell;
 } rates_t;
 
 extern rates_t btcrates;
